Break out the `console` command
Continuing with the clean-up of `pkg/cmd/pulumi`, this factors out the
`console` command.
