+++ pkg/cmd/pulumi/console/console.go
@@ -1,4 +1,4 @@
-// Copyright 2016-2023, Pulumi Corporation.
+// Copyright 2016-2024, Pulumi Corporation.
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package main
+package console
 
 import (
 	"errors"
@@ -32,7 +32,7 @@ import (
 	"github.com/pulumi/pulumi/sdk/v3/go/common/workspace"
 )
 
-func newConsoleCmd() *cobra.Command {
+func NewConsoleCmd() *cobra.Command {
 	var stackName string
 	cmd := &cobra.Command{
 		Use:   "console",
